docs: add link to download Chinese package of ocwr
A soft workaround for users in China unable to bypass the Great
Firewall (GFW). Likely to become a permanent workaround.

The scripts are preconfigured to install WinRAR in Simplified Chinese.
If the user requires the installation to be in Traditional Chinese,
they will have to follow the documentation in order to install their
preferred language.