docs: add link to download Chinese package of ocwr
A soft workaround for users in China unable to bypass the Great Firewall. Likely to become a permanent workaround.

Henceforth in this commit, "WinRAR 7.10" will be referred to as "the latest release"; and "WinRAR 7.01" will be referred to as "the Chinese version" and "Chinese version". No other implications are to be made therein.

There is currently no Chinese version available for the latest release of WinRAR; therefore, in order to allow users unable to bypass the GFW, this package was created to contain the most recent Chinese version for a better experience.

However, for Chinese users who are able to bypass the GFW, the scripts which are a part of ocwr which are able to download installers from the internet (specifically `oneclickrar.cmd` and `installrar.cmd`) have been pre-configured to download and install the most recent Chinese version.

The Chinese version will be installed in Simplified Chinese. If the user requires the installation to be in Traditional Chinese, they will have to follow the documentation in order to install their preferred language.