fix: download-only failure
Download-only fails when a WinRAR version is not specified.

`CUSTOM_DOWNLOAD` was set whenever download-only is enabled, resulting
in the script requesting a specific version of WinRAR to download,
however, the script is expected to always download the default `LATEST`
version of WinRAR if no version is specified.

Removing this flag allows the script to proceed with the expected
behavior.