refactor: remove termination scriptblocks
This commit removes the option to run commands within a scriptblock for
`Stop-OcwrOperation` function.

Depending on whether the commands were run as a background process with
the '&' operator or plainly within the scriptblock, the command would
result in errors such as "The expression after '&' in a pipeline element
produced an object that was not valid. It must result in a command name,
a script block, or a CommandInfo object." or similar.

As a result, if/until this is rectified, the commands would hereafter
be run before the function and then call to terminate.