change '-' to 'en' in language list
The position in the list cannot be removed for sake of the
`Get-LanguageName` logic and 'en' is more immediately understandable
than a '-'.

If an unknown language was provided, the TAGS variable would be assigned
a null value which will be later interpreted as the default English
language. Therefore providing a language 'en'—whether or not it's in the
language list—has has the same result.