feat(ps1): grab latest version from winrar website
installrar.ps1 and oneclickrar.ps1 are now able to retrieve the latest
version of WinRAR by scraping the What's New page on the webiste.

If the scraping fails, the previous feature to check for updates
manually in commit a966a48520a7281d082a939223ef2c66a6cf39fa will serve
as a backup for this.