fix: beta version download bug
The download TAGS would be set to null when checking to ensure that the
language code 'en' for English is not included as an option. This
inadvertently, removed any other TAGS that were present in the variable,
i.e. the beta tags.

For now the line is commented out as the current workaround. Users will
have to be smart enough on their own to not include the English tags in
their config since it's the default anyway.